face: refactoring of factory methods
