+++ modules/face/samples/facerec_save_load.cpp
@@ -127,7 +127,7 @@ int main(int argc, const char *argv[]) {
     //
     //      cv::createEigenFaceRecognizer(0, 123.0);
     //
-    Ptr<BasicFaceRecognizer> model0 = createEigenFaceRecognizer();
+    Ptr<EigenFaceRecognizer> model0 = EigenFaceRecognizer::create();
     model0->train(images, labels);
     // save the model to eigenfaces_at.yaml
     model0->save("eigenfaces_at.yml");
@@ -135,8 +135,7 @@ int main(int argc, const char *argv[]) {
     //
     // Now create a new Eigenfaces Recognizer
     //
-    Ptr<BasicFaceRecognizer> model1 = createEigenFaceRecognizer();
-    model1->load("eigenfaces_at.yml");
+    Ptr<EigenFaceRecognizer> model1 = Algorithm::load<EigenFaceRecognizer>("eigenfaces_at.yml");
     // The following line predicts the label of a given
     // test image:
     int predictedLabel = model1->predict(testSample);
