+++ modules/face/src/eigen_faces.cpp
@@ -16,7 +16,8 @@
  *   See <http://www.opensource.org/licenses/bsd-license>
  */
 #include "precomp.hpp"
-#include "face_basic.hpp"
+#include <opencv2/face.hpp>
+#include "face_utils.hpp"
 #include <set>
 #include <limits>
 #include <iostream>
@@ -28,14 +29,17 @@ namespace face
 
 // Turk, M., and Pentland, A. "Eigenfaces for recognition.". Journal of
 // Cognitive Neuroscience 3 (1991), 71–86.
-class Eigenfaces : public BasicFaceRecognizerImpl
+class Eigenfaces : public EigenFaceRecognizer
 {
 
 public:
     // Initializes an empty Eigenfaces model.
     Eigenfaces(int num_components = 0, double threshold = DBL_MAX)
-        : BasicFaceRecognizerImpl(num_components, threshold)
-    {}
+        //: BasicFaceRecognizerImpl(num_components, threshold)
+    {
+        _num_components = num_components;
+        _threshold = threshold;
+    }
 
     // Computes an Eigenfaces model with images in src and corresponding labels
     // in labels.
@@ -122,7 +126,7 @@ void Eigenfaces::predict(InputArray _src, Ptr<PredictCollector> collector) const
     }
 }
 
-Ptr<BasicFaceRecognizer> createEigenFaceRecognizer(int num_components, double threshold)
+Ptr<EigenFaceRecognizer> EigenFaceRecognizer::create(int num_components, double threshold)
 {
     return makePtr<Eigenfaces>(num_components, threshold);
 }
