+++ modules/face/src/lbph_faces.cpp
@@ -17,7 +17,7 @@
  */
 #include "precomp.hpp"
 #include "opencv2/face.hpp"
-#include "face_basic.hpp"
+#include "face_utils.hpp"
 
 namespace cv { namespace face {
 
@@ -46,8 +46,8 @@ private:
 
 
 public:
-    using FaceRecognizer::save;
-    using FaceRecognizer::load;
+    using FaceRecognizer::read;
+    using FaceRecognizer::write;
 
     // Initializes this LBPH Model. The current implementation is rather fixed
     // as it uses the Extended Local Binary Patterns per default.
@@ -94,11 +94,11 @@ public:
     // Send all predict results to caller side for custom result handling
     void predict(InputArray src, Ptr<PredictCollector> collector) const;
 
-    // See FaceRecognizer::load.
-    void load(const FileStorage& fs);
+    // See FaceRecognizer::write.
+    void read(const FileNode& fn);
 
     // See FaceRecognizer::save.
-    void save(FileStorage& fs) const;
+    void write(FileStorage& fs) const;
 
     CV_IMPL_PROPERTY(int, GridX, _grid_x)
     CV_IMPL_PROPERTY(int, GridY, _grid_y)
@@ -110,7 +110,7 @@ public:
 };
 
 
-void LBPH::load(const FileStorage& fs) {
+void LBPH::read(const FileNode& fs) {
     fs["radius"] >> _radius;
     fs["neighbors"] >> _neighbors;
     fs["grid_x"] >> _grid_x;
@@ -132,7 +132,7 @@ void LBPH::load(const FileStorage& fs) {
 }
 
 // See FaceRecognizer::save.
-void LBPH::save(FileStorage& fs) const {
+void LBPH::write(FileStorage& fs) const {
     fs << "radius" << _radius;
     fs << "neighbors" << _neighbors;
     fs << "grid_x" << _grid_x;
@@ -407,11 +407,10 @@ void LBPH::predict(InputArray _src, Ptr<PredictCollector> collector) const {
     }
 }
 
-Ptr<LBPHFaceRecognizer> createLBPHFaceRecognizer(int radius, int neighbors,
+Ptr<LBPHFaceRecognizer> LBPHFaceRecognizer::create(int radius, int neighbors,
                                              int grid_x, int grid_y, double threshold)
 {
     return makePtr<LBPH>(radius, neighbors, grid_x, grid_y, threshold);
 }
-
 
 }}
