+++ modules/face/src/precomp.hpp
@@ -49,10 +49,8 @@
 #include "opencv2/core/private.hpp"
 #include "opencv2/core/persistence.hpp"
 #include <map>
-#include <iostream>
 #include <set>
 #include <limits>
-#include <iostream>
 
 #endif
